@@ -22,9 +22,10 @@
   },
   "devDependencies": {
     "@intlify/vue-i18n-loader": "^3.0.0",
+    "@types/node": "^20.8.10",
     "@types/vue-router": "^2.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
     "@vue/cli-plugin-babel": "~5.0.8",
     "@vue/cli-plugin-eslint": "~5.0.8",
     "@vue/cli-plugin-typescript": "~5.0.0",
@@ -32,7 +33,7 @@
     "@vue/compiler-sfc": "^3.2.47",
     "@vue/eslint-config-typescript": "^9.1.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^8.35.0",
+    "eslint": "^8.53.0",
     "eslint-config-prettier": "^8.6.0",
     "prettier": "2.8.4",
     "typescript": "~4.5.5",
