@@ -4,7 +4,9 @@
       <table class="w-full">
         <tr>
           <td class="w-2">
-            <label class="form-label" for="isbn">{{ $t("isbn") }}</label>
+            <label class="form-label" for="isbn">{{
+              i18n.global.t("isbn")
+            }}</label>
           </td>
           <td class="w-6">
             <IsbnSearch
@@ -33,26 +35,28 @@
             <InputText
               id="title"
               v-model="bookCopy.title"
+              class="w-full"
               name="title"
               type="text"
-              class="w-full"
             />
           </td>
         </tr>
         <tr>
           <td>
-            <label class="form-label" for="editor">{{ $t("editor") }}</label>
+            <label class="form-label" for="editor">{{
+              i18n.global.t("editor")
+            }}</label>
           </td>
           <td>
             <DropDown
               id="editor"
               v-model="bookCopy.editor"
-              :editable="true"
               :filter="true"
               :options="editors"
-              :placeholder="$t('select')"
-              option-label="name"
+              :placeholder="i18n.global.t('select')"
               class="w-full"
+              editable
+              option-label="name"
               @change="onEditorChange"
             />
           </td>
@@ -60,7 +64,9 @@
 
         <tr>
           <td style="vertical-align: top">
-            <label class="form-label" for="authors">{{ $t("authors") }}</label>
+            <label class="form-label" for="authors">{{
+              i18n.global.t("authors")
+            }}</label>
           </td>
           <td>
             <AuthorForm
@@ -75,7 +81,7 @@
         <tr>
           <td>
             <label class="form-label" for="distributor">{{
-              $t("distributor")
+              i18n.global.t("distributor")
             }}</label>
           </td>
           <td>
@@ -85,7 +91,7 @@
               :editable="true"
               :filter="true"
               :options="distributors"
-              :placeholder="$t('select')"
+              :placeholder="i18n.global.t('select')"
             />
           </td>
         </tr>
@@ -100,35 +106,41 @@
         </tr>
         <tr>
           <td>
-            <label class="form-label" for="price">{{ $t("price") }}</label>
+            <label class="form-label" for="price">{{
+              i18n.global.t("price")
+            }}</label>
           </td>
           <td>
             <InputNumber
               v-model="bookCopy.price"
+              class="w-full"
               currency="CHF"
               mode="currency"
-              class="w-full"
             />
           </td>
         </tr>
 
         <tr>
           <td>
-            <label class="form-label" for="amount">{{ $t("quantity") }}</label>
+            <label class="form-label" for="amount">{{
+              i18n.global.t("quantity")
+            }}</label>
           </td>
           <td>
             <InputNumber
               id="amount"
               v-model="bookCopy.amount"
               :show-buttons="false"
-              name="amount"
               class="w-full"
+              name="amount"
             />
           </td>
         </tr>
         <tr>
           <td>
-            <label class="form-label" for="tags">{{ $t("tags") }}</label>
+            <label class="form-label" for="tags">{{
+              i18n.global.t("tags")
+            }}</label>
           </td>
           <td class="p-fluid">
             <AutoComplete
@@ -163,17 +175,20 @@
     </form>
   </div>
 </template>
-<script setup>
+<script lang="ts" setup>
 import AuthorForm from "@/components/AuthorForm";
 import StockService from "@/service/StockService";
 import IsbnSearch from "@/components/IsbnSearch";
-import { defineEmits, defineProps, inject, ref } from "vue";
+import { defineEmits, defineProps, inject, ref, Ref } from "vue";
 import { useFetchDistributors, useFetchEditors } from "@/composables/useFetch";
 import i18n from "@/i18n";
 import tagService from "@/service/TagService";
+import { ScraperBook } from "@/models/ScraperBook";
+import { Emitter } from "mitt";
+import {MessageEvent} from "@/classes/Message";
 
 const props = defineProps({
-  book: {},
+  book: ScraperBook,
   editMode: {
     type: Boolean,
     default: false,
@@ -191,11 +206,13 @@ bookCopy.value.authors = bookCopy.value.authors.map((a) => {
 let editors = useFetchEditors();
 let distributors = useFetchDistributors();
 
-const tags = ref([]);
-const tagsSuggestions = ref([]);
+const tags = ref<TagDto[]>([]);
+const tagsSuggestions = ref<TagDto[]>([]);
+const emitter: Emitter<MessageEvent> = inject("emitter") as Emitter<MessageEvent>
+
 loadTags();
 
-function loadTags() {
+function loadTags(): void {
   console.debug("loading tags");
   tagService
     .getTags()
@@ -203,34 +220,34 @@ function loadTags() {
       tags.value = data;
     })
     .catch((e) =>
-      emitter.emit("notify", {
-        severity: "error",
-        content: i18n.global.t("couldNotLoadTagsMessage", { error: e }),
-      })
+      emitter.emit(
+        "notify",
+          {severity: "error", content: i18n.global.t("couldNotLoadTagsMessage", { error: e })})
     );
 }
 
-const emitter = inject("emitter");
 
-const changeEditor = (editorName) => {
+function changeEditor(editorName: string): void {
+  console.debug("editor name");
   console.log(editorName);
   const editor = editors.value.find((e) => e.name === editorName);
   if (editor && editor.defaultDistributor !== "null") {
     bookCopy.value.distributor = editor.defaultDistributor;
   }
   bookCopy.value.editor = editorName;
-};
+}
+
 if (bookCopy.value.editor && bookCopy.value.editor.length > 0) {
   changeEditor(bookCopy.value.editor);
 }
 
 // let formData = ref(null);
 
-let preventSubmit = ref(false);
+let preventSubmit = ref<boolean>(false);
 
 const emit = defineEmits(["close-dialog"]);
 
-const fillData = (data) => {
+function fillData(data): void {
   console.debug("found data");
   console.debug(data);
   bookCopy.value.isbn = data.isbn;
@@ -239,32 +256,30 @@ const fillData = (data) => {
     return { value: a };
   });
   bookCopy.value.editor = data.editor;
+  bookCopy.value.distributor = data.distributor;
   bookCopy.value.description = data.description;
-  if (bookCopy.value.editor && bookCopy.value.editor.length > 0) {
-    changeEditor(bookCopy.value.editor);
-  }
   bookCopy.value.price = data.price;
   bookCopy.value.coverUrl = data.coverUrl;
   bookCopy.value.hasCover = data.hasCover;
   bookCopy.value.tags = data.tags;
   console.debug("data filled");
-};
+}
 
-const enableSubmit = (enable) => {
-  console.debug("in prevent submit" + enable)
+function enableSubmit(enable: boolean): void {
+  console.debug("in prevent submit" + enable);
   preventSubmit.value = enable;
-};
+}
 
-const addAuthor = () => {
+function addAuthor(): void {
   bookCopy.value.authors.push({ value: "" });
-};
+}
 
-const deleteAuthor = (index) => {
+function deleteAuthor(index: number): void {
   bookCopy.value.authors.splice(index, 1);
-};
+}
 
 // needed because primevue doesn't understand that it can take the value itself
-const processForm = () => {
+function processForm(): void {
   if (preventSubmit.value) {
     return;
   }
@@ -282,15 +297,17 @@ const processForm = () => {
         content: i18n.global.t("stockNotUpdatableMessage"),
       })
     );
-};
+}
 
-const onEditorChange = (event) => {
-  if (event.value && event.value.name !== null) {
+function onEditorChange(event: Ref) {
+  console.debug("type of event");
+  console.debug(event);
+  if (event.value && event.value.name !== undefined) {
     changeEditor(event.value.name);
   }
-};
+}
 
-const fileUpload = (event) => {
+function fileUpload(event: Event): void {
   const formData = new FormData();
   formData.append("file", event.files[0]);
   StockService.addCover(formData, bookCopy.value.isbn)
@@ -306,22 +323,26 @@ const fileUpload = (event) => {
         content: i18n.global.t("coulntUploadPictureMessage"),
       })
     );
-};
+}
 
-const bookContainTag = (tag) => {
+function bookContainTag(tag: TagDto): boolean {
   for (const t of bookCopy.value.tags) {
     if (t.id === tag.id) {
-      return true
+      return true;
     }
   }
-  return false
+  return false;
 }
-const searchTags = (event) => {
+
+function searchTags(event: Event): void {
   console.debug(event.query);
   tagsSuggestions.value = tags.value.filter((tag) => {
-    return tag.name.toLowerCase().startsWith(event.query.toLowerCase()) && !bookContainTag(tag);
+    return (
+      tag.name.toLowerCase().startsWith(event.query.toLowerCase()) &&
+      !bookContainTag(tag)
+    );
   });
-};
+}
 </script>
 
 <style scoped>
@@ -335,5 +356,4 @@ div :deep(.p-hidden-accessible) {
   position: absolute;
   width: 1px;
 }
-
 </style>
