@@ -1,17 +1,18 @@
-import {ref} from "vue"
+import {ref, Ref} from "vue"
 import EditorService from "@/service/EditorService"
 import DistributorService from "@/service/DistributorService"
 import SaleService from "@/service/SaleService";
 import i18n from "@/i18n";
 import mitt from "mitt";
+import {SaleList} from "@/models/SaleDto";
 
 const emitter = mitt()
 
-export function useFetchEditors() {
-    const editors = ref([])
+export function useFetchEditors() : Ref<Editor[]> {
+    const editors = ref<Editor[]>([])
     EditorService.getAll()
-        .then((response) => response.json())
-        .then((data) => {
+        .then((response: Response) => response.json())
+        .then((data: Editor[]) => {
             console.log('fetched editors')
             // needded because Drodpown doesn't work with simple list
             editors.value = data
@@ -23,11 +24,11 @@ export function useFetchEditors() {
     return editors
 }
 
-export function useFetchDistributors() {
-    const distributors = ref([])
+export function useFetchDistributors() : Ref<string[]> {
+    const distributors = ref<string[]>([])
     DistributorService.getAll()
-        .then((response) => response.json())
-        .then((data) => {
+        .then((response: Response) => response.json())
+        .then((data: string[]) => {
             distributors.value = data
             console.log(data)
         })
@@ -38,11 +39,11 @@ export function useFetchDistributors() {
     return distributors
 }
 
-export function useFetchInvoices(startTime: Date, endTime: Date) {
-    const invoices = ref([])
+export function useFetchInvoices(startTime: Date, endTime: Date) : Ref<SaleList[]> {
+    const invoices = ref<SaleList[]>([])
     SaleService.getSales(startTime, endTime)
-        .then((response) => response.json())
-        .then((data) => {
+        .then((response: Response) => response.json())
+        .then((data: SaleList[]) => {
             invoices.value = data
             console.debug(data)
         })
