@@ -1,13 +1,13 @@
-import {ref} from "vue"
+import {ref,Ref} from "vue"
 
-export function useSearch(list: any, event: any) {
+export function useSearch(list: never[], event: any) : Ref<string[]> {
     console.log("use search")
-    const filteredResult: any = ref([])
+    const filteredResult = ref<string[]>([])
     console.log(list)
     if (!event.query.trim().length) {
         filteredResult.value = [...list]
     } else {
-        list.forEach((e: string) => {
+        list.forEach((e: string) : void => {
           if (e.toLowerCase().includes(event.query.toLowerCase())) {
             filteredResult.value.push(e);
           }
