@@ -1,4 +1,4 @@
-class BasketRow {
+export class BasketRow {
     isbn!: string;
     title!: string;
     quantity!: number;
