@@ -1,10 +1,9 @@
 import handleErrors from "@/helpers/httpHelper";
 
 class DistributorService {
-    getAll() {
-        return fetch("/api/distributor")
-            .then(handleErrors)
-    }
+  getAll(): Promise<Response> {
+    return fetch("/api/distributor").then(handleErrors);
+  }
 }
 
 export default new DistributorService();
