@@ -1,10 +1,9 @@
 import handleErrors from "@/helpers/httpHelper";
 
 class EditorService {
-    getAll() {
-        return fetch("/api/editor")
-            .then(handleErrors)
-    }
+  getAll(): Promise<Response> {
+    return fetch("/api/editor").then(handleErrors);
+  }
 }
 
 export default new EditorService();
