@@ -1,19 +1,21 @@
 import handleErrors from "@/helpers/httpHelper";
 
 class SaleService {
+  getSales(startTime: Date, endTime: Date): Promise<Response> {
+    console.debug(startTime);
+    console.debug(endTime);
+    return fetch(
+      "/api/sales?" +
+        new URLSearchParams({
+          startTime: startTime.toISOString(),
+          endTime: endTime.toISOString(),
+        })
+    ).then(handleErrors);
+  }
 
-    getSales(startTime: Date, endTime: Date) {
-        console.debug(startTime)
-        console.debug(endTime)
-        return fetch("/api/sales?" + new URLSearchParams({startTime: startTime.toISOString(), endTime: endTime.toISOString()}))
-            .then(handleErrors)
-    }
-
-    getInvoice(invoiceId: number) {
-        return fetch("api/invoices/" + invoiceId.toString())
-            .then(handleErrors)
-    }
-
+  getInvoice(invoiceId: number): Promise<Response> {
+    return fetch("api/invoices/" + invoiceId.toString()).then(handleErrors);
+  }
 }
 
-export default new SaleService()
+export default new SaleService();
