@@ -1,71 +1,53 @@
 <template>
-  <h1>{{ $t('addBook') }}</h1>
+  <h1>{{ i18n.global.t("addBook") }}</h1>
   <StockEntryEdit
-      :key="reloadCount"
-      v-model:book="initBook"
-      :process-form-function="addBook"
-      class="w-64rem"
+    :key="reloadCount"
+    v-model:book="initBook"
+    :process-form-function="addBook"
+    class="w-64rem"
   />
-
 </template>
 
-<script>
+<script lang="ts" setup>
 import StockEntryEdit from "@/components/StockEntryEdit";
 import StockService from "@/service/StockService";
 import i18n from "@/i18n";
-import {inject} from "vue";
+import { inject, ref } from "vue";
+import { ScraperBook } from "@/models/ScraperBook";
+import { Emitter } from "mitt"
+import {MessageEvent} from "@/classes/Message";
 
-export default {
-  el: '#addbook',
-  name: "AddBook",
-  components: {StockEntryEdit},
-  data() {
-    return {
-      reloadCount: 0,
-      initBook: {
-        isbn: null,
-        title: null,
-        authors: [],
-        editor: null,
-        distributor: null,
-        description: null,
-        price: null,
-        coverUrl: null,
-        hasCover: null,
-        amount: 1,
-        tags: [],
-      },
-      emitter: inject('emitter')
-    }
-  },
-  methods: {
-    addBook: function (book) {
-      console.debug("addBook.addbook")
-      const body = {
-        isbn: book.isbn,
-        title: book.title,
-        // authors: null,
-        authors: book.authors.map(a => a.value),
-        editor: book.editor,
-        distributor: book.distributor,
-        description: book.description,
-        price: book.price,
-        coverUrl: book.coverUrl,
-        hasCover: book.hasCover,
-        amount: book.amount,
-        tags: book.tags,
-      }
-      this.reloadCount++
-      this.emitter.emit('notify',{severity: 'success', content: i18n.global.t('stockHasBeenModifiedMessage')})
-      return StockService.addBook(body)
-    },
-  }
+const reloadCount = ref(0);
+const initBook = ref(new ScraperBook());
+initBook.value.amount = 1;
 
+const emitter: Emitter<MessageEvent> = inject("emitter") as Emitter<MessageEvent>
+
+function addBook(book: ScraperBook) {
+  const body = {
+    isbn: book.isbn,
+    title: book.title,
+    // authors: null,
+    authors: book.authors.map((a) => a.value),
+    editor: book.editor,
+    distributor: book.distributor,
+    description: book.description,
+    price: book.price,
+    coverUrl: book.coverUrl,
+    hasCover: book.hasCover,
+    amount: book.amount,
+    tags: book.tags,
+  };
+  reloadCount.value++;
+  emitter.emit("notify", {
+    severity: "success",
+    content: i18n.global.t("stockHasBeenModifiedMessage")
+  });
+  return StockService.addBook(body);
 }
 </script>
 
 <style scoped>
-
 div :deep(.p-hidden-accessible) {
   border: 0;
   clip: rect(0 0 0 0);
@@ -76,5 +58,4 @@ div :deep(.p-hidden-accessible) {
   position: absolute;
   width: 1px;
 }
-
 </style>
